Nicholas/Finish Change status order, make controller, and service...

See merge request AdvProg/reguler-2023/mahasiswa/kelas-a/2106629963-Muflih-Naufal-Maxi-/kang-warnet/cafe-service!16

+++ src/main/java/id/ac/ui/cs/advprog/cafeservice/service/MenuItemServiceImpl.java
@@ -50,6 +50,8 @@ public class MenuItemServiceImpl implements MenuItemService {
 
     @Override
     public void delete(String id) {
+        Optional<MenuItem> menuItem = menuItemRepository.findById(id);
+        if (menuItem.isEmpty()) throw new MenuItemDoesNotExistException(id);
         menuItemRepository.deleteById(id);
     }
